Bump about library to 8.9.4
The latest 10.0.0 has latest AndroidX dependencies that need targetSdk
to 31. Looks like broker needs to keep to lower target SDK, so this CL
skip latest about library, and bump it to one newer version but works
fine with current dependencies.

Testing it with installable_broker, and about page looks great.

Signed-off-by: utzcoz <utzcoz@outlook.com>